app/requirements.txt
alabaster==0.7.13
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-debug-toolbar==4.1.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
emoji==2.6.0
emoji-data-python==1.5.0
factory-boy==3.2.1
Faker==19.1.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zope.interface==6.0
